requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cognitive_complexity==1.3.0
coverage==7.5.3
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-environ==0.11.2
django-filter==24.3
django-object-actions==4.3.0
django-split-settings==1.3.1
django-timezone-field==7.0
djangorestframework==3.15.2
eradicate==2.3.0
flake8==7.0.0
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-print==5.0.0
flake8_cognitive_complexity==0.1.0
idna==3.9
iniconfig==2.0.0
isort==5.13.2
kombu==5.4.1
loguru==0.7.2
Markdown==3.7
mccabe==0.7.0
packaging==24.0
pluggy==1.5.0
progressbar2==4.4.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
pyTelegramBotAPI==4.18.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.2
redis==5.0.8
requests==2.32.3
requests-mock==1.12.1
setuptools==70.0.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.5.1
telebot==0.0.5
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
